1) Wrote the shared resolver daemon. 2) Changed hostdb_count in netresolve.c from int to long.

+++ proto2/src/inc/netresolve.h
@@ -26,7 +26,7 @@ struct huinfo { /* host/user information */
 
 extern struct hostinfo *hostdb;
 extern struct husrinfo *userdb;
-extern unsigned int hostdb_count;
+extern unsigned long hostdb_count;
 
 #ifdef SPAWN_HOST_RESOLVER
 extern void kill_resolver(void);
